### 2019-02-28 #1290 @whedon
[PRE REVIEW]: healpy: equal area pixelization and spherical harmonics transforms for data on the sphere in Python
**Submitting author:** @zonca (<a href="http://orcid.org/0000-0001-6841-1058">Andrea Zonca</a>)
**Repository:** <a href="https://github.com/healpy/healpy" target ="_blank">https://github.com/healpy/healpy</a>
**Version:** 1.12.8
