### 2019-02-28 #1290 @danielskatz
👋 @zonca - in your paper.md, all authors should have affiliations.  And if there is no ORCID, please remove the ```orcid:``` line.  After fixing these, try to recompile by entering ```@whedon generate pdf``` as a new comment in this issue
