### 2019-02-28 503b0e4
JOSS paper, add @lpsinger affiliations, remove empty orcid
