### 2019-02-28 c75b7d7
JOSS paper, fix bib formatting issue
