+++ paper/paper.bib
@@ -72,9 +72,8 @@ archivePrefix = "arXiv",
 
 
 @article{planck18,
-      author         = "Planck Collaboration"
-      title          = "{Planck 2018 results. I. Overview and the cosmological
-                        legacy of Planck}",
+      author         = "Planck Collaboration",
+      title          = "{Planck 2018 results. I. Overview and the cosmological legacy of Planck}",
       collaboration  = "Planck",
       year           = "2018",
       eprint         = "1807.06205",
