### 2019-02-28 eda086b
JOSS paper, set figure in own line
