### 2019-02-28 #1290 @danielskatz
Please also note https://github.com/healpy/healpy/pull/537
