+++ paper/paper.bib
@@ -72,7 +72,7 @@ archivePrefix = "arXiv",
 
 
 @article{planck18,
-      author         = "Planck Collaboration",
+      author         = "{Planck Collaboration}",
       title          = "{Planck 2018 results. I. Overview and the cosmological legacy of Planck}",
       collaboration  = "Planck",
       year           = "2018",
