### 2019-03-04 #1298 @whedon
[REVIEW]: healpy: equal area pixelization and spherical harmonics transforms for data on the sphere in Python
**Submitting author:** @zonca (<a href="http://orcid.org/0000-0001-6841-1058">Andrea Zonca</a>)
**Repository:** <a href="https://github.com/healpy/healpy" target ="_blank">https://github.com/healpy/healpy</a>
**Version:** v1.12.9
