### 2019-03-04 #1290 @xuanxu
@harpolea, please head over to #1298 for the actual review
