### 2019-03-06 healpy/healpy#538 @harpolea
Tutorial file location error
In the section 'Reading and writing maps to file' of the tutorial, it says to read a map from the file `../healpy/test/data/wmap_band_imap_r9_7yr_W_v4.fits`. Before this part, there should be an instruction to move to the healpy directory (or somewhere where this relative path will work), or the file location should be changed to an absolute path.

I also cannot find that exact file - using `wmap_band_iqumap_r9_7yr_W_v4_udgraded32.fits` instead does not give me the same output as shown in the Visualization section of the tutorial (it looks like it's lower resolution?). The same goes for the file in the section on masked maps - I think this should be `wmap_temperature_analysis_mask_r9_7yr_v4_udgraded32.fits`.
