### 2019-03-06 healpy/healpy#539 @harpolea
Normalized CMB spectrum in tutorial
Working through the tutorial, I was not able to replicate the plot of the normalized CMB spectrum at the end. The plot that I generated was:
![cmb](https://user-images.githubusercontent.com/8247792/53896325-c96cb380-4001-11e9-904b-bdcf36c98879.png)
It looks like the values of `cl` returned by the `hp.anafast` function when I tried it were different from the values used in the tutorial example?
