### 2019-03-06 #1298 @harpolea
Hi all,
Overall, this package looks great and I've ticked off all the items in the review checklist. However I have a few additional points/suggestions:

- I worked through the tutorial and found a couple of bugs. I filed these as issues (healpy/healpy#538 and healpy/healpy#539)
- I checked the test coverage and it is a little low for some modules (< 60%) - it would be great if this could be improved
- A minor point: it would be nice to have a description/overview of the project in the documentation (i.e. not just in the README)
- Checking the contributors to the project, there are a few people who appear to have made substantial contributions (> 500 additions) but who are not listed in the author list. The criteria for who is listed as an author or is considered as a 'core developer' is not given, so I was wondering how this was determined?
