+++ doc/index.rst
@@ -1,7 +1,3 @@
-.. healpy documentation master file, created by sphinx-quickstart on Fri Nov 21 13:06:27 2008.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
 Welcome to the healpy documentation
 ===================================
 
