### 2019-03-06 healpy/healpy#538 @DanielLenz
Thanks for pointing this out! Interestingly, I also couldn't find the file in the history, the following returns nothing:

    git log --all --full-history -- healpy/test/data/wmap_band_imap_r9_7yr_W_v4.fits

It appears though that the smaller, lower resolution files were added in 2014:

    git log --all --full-history -- healpy/test/data/wmap_band_iqumap_r9_7yr_W_v4_udgraded32.fits
```
commit 15d95e6a87a71e3c605c2e8925860dfe7cdb141c
Author: Andrea Zonca <code@andreazonca.com>
Date:   Fri Jan 10 10:30:57 2014 -0800

    Refactoring of the Spherical Armonics tests

    Now nside32 test maps are included in the repository
```

This also explains the issue described in https://github.com/healpy/healpy/issues/539. Estimating the power spectrum up to lmax=1024 with a map at nside=32 won't work, a conservative upper limit for lmax is 2*nside.

My guess would be that the higher resolution files were remove because of their size (~40MB for IQU at nside512). @zonca , what do you think? Add the higher resolution WMAP data again, or update the tutorial with the nside32 map/simulated data?
