### 2019-03-06 #1298 @zonca
thanks @harpolea !
the tutorial definitely needs some love, @daniellenz would you like to work on that? I think at some point we lost the script to download the high resolution WMAP maps, I recovered it from an ancient healpy version, see https://github.com/healpy/healpy/blob/master/healpy/test/data/get_wmap_maps.sh
We should tell people to run that before running the tutorial so they have some real data, or just rework the tutorial at lower resolution, whatever you think is best.
