+++ .readthedocs.yml
@@ -1,6 +1,5 @@
 build:
   image: latest
 
-python:
-  version: 3.6
-  setup_py_install: true
+conda:
+  file: .rtd-environment.yml
