+++ .rtd-environment.yml
@@ -1,0 +1,16 @@
+name: healpy
+
+channels:
+    - conda-forge
+
+dependencies:
+    - python>=3.6
+    - astropy
+    - Cython
+    - matplotlib
+    - numpy
+    - gcc_linux-64
+    - gxx_linux-64
+    - autoconf
+    - automake
+    - libtool
