### 2019-03-06 158d081
[docs] install pkg-config on readthedocs
