### 2019-03-06 869afa7
[docs] install numpydoc on readthedocs
