+++ .rtd-environment.yml
@@ -15,3 +15,4 @@ dependencies:
     - automake
     - libtool
     - pkg-config
+    - numpydoc
