### 2019-03-06 7784e33
test that rotator raises exception if coord and rot are incompatible
