### 2019-03-06 d1b2223
test rotator equal and inverse
