### 2019-03-06 2616496
test euler_matrix_new
