### 2019-03-07 #1298 @DanielLenz
Thanks @harpolea  for the review, these are all great comments!

Thanks @zonca for adding the wmap download script back to the repo. I'd be happy to update the tutorial, and I think the higher resolution data are important to have (e.g. Cl estimation doesn't really make sense on nside=32 data).
