### 2019-03-07 #1298 @zonca
@harpolea current status:

* [ ] **tutorial**, @DanielLenz working on it
* [ ] **coverage**, have pull request ready (https://github.com/healpy/healpy/pull/541) to bring coverage over 80% for all the computational functions. I would prefer not to add more test on the modules related to visualization
* [x] **intro in docs**, I discovered our webhook was not working anymore and the docs were 2 months old, fixed now: https://healpy.readthedocs.io/en/latest/
* [ ] **contributors** authors are the core devs, (+ the HEALPix team as they maintain the upstream package and coordinate joint releases) and the core devs are people who monitor and answer the support tickets, anyone can become a core dev just asking me or another core dev. @DanielLenz is the most recent core dev and has relatively few commits compared to other more random contributors, but he is now the person spending more time on the project and this is probably going to increase in the future.
