### 2019-03-08 edbad32
How to update the HEALPix C++ library (#535)
Also includes a policy discussed with the HEALPix team
