### 2019-03-12 healpy/healpy#538 @zonca
@DanielLenz any update on fixing the tutorial?
