### 2019-03-13 healpy/healpy#538 @DanielLenz
Haven't had time yet, will do it before the weekend.
