### 2019-03-14 f0b365b
Support pixel weights in smoothing
