### 2019-03-15 c827eed
Add download instructions for high-res WMAP data
