### 2019-03-15 b6f87d3
Add lmax to hp.pixwin. Black formatting
