### 2019-03-15 6f247c6
Add tests for hp.pixwin()
