+++ healpy/test/test_fitsfunc.py
@@ -14,7 +14,6 @@ from ..sphtfunc import *
 
 
 class TestFitsFunc(unittest.TestCase):
-
     def setUp(self):
         self.nside = 512
         self.m = np.arange(healpy.nside2npix(self.nside))
@@ -159,7 +158,6 @@ class TestFitsFunc(unittest.TestCase):
 
 
 class TestFitsFuncGzip(unittest.TestCase):
-
     def setUp(self):
         self.nside = 4
         self.m = np.arange(healpy.nside2npix(self.nside))
@@ -179,7 +177,6 @@ class TestFitsFuncGzip(unittest.TestCase):
 
 
 class TestReadWriteAlm(unittest.TestCase):
-
     def setUp(self):
 
         s = Alm.getsize(256)
@@ -256,7 +253,6 @@ class TestReadWriteAlm(unittest.TestCase):
 
 
 class TestReadWriteCl(unittest.TestCase):
-
     def setUp(self):
         self.filename = "test_cl.fits"
 
