### 2019-03-15 healpy/healpy#538 @zonca
fixed in #543
