### 2019-03-15 healpy/healpy#539 @zonca
fixed in #543
