### 2019-03-15 1fb3783
Update tests for off-by-one error
