### 2019-03-16 d1e9ac4
Pixwin lmax (#544)
* Add lmax to hp.pixwin. Black formatting

* Add tests for hp.pixwin()

* Black formatting

* Fix off-by-one error

* Update tests for off-by-one error

* Remove comment
