+++ healpy/__init__.py
@@ -100,11 +100,4 @@ from .visufunc import (
     azeqview,
 )
 from .zoomtool import mollzoom, set_g_clim
-from .fitsfunc import (
-    write_map,
-    read_map,
-    read_alm,
-    write_alm,
-    write_cl,
-    read_cl,
-)
+from .fitsfunc import write_map, read_map, read_alm, write_alm, write_cl, read_cl
