### 2019-03-18 #1298 @zonca
Hi @harpolea, we just merged a fix for the issue on the tutorial
