### 2019-03-19 #1298 @harpolea
Great! I reran the tutorial and everything looks good to me.

One small thing: I did try installing from conda on a different machine when I ran it and it didn't install the most recent version (e.g. it didn't include the new script to download the high res data), so I think that needs to be updated?
