### 2019-03-21 3b65363
use data.conf.set_temp to set url and timeout
