### 2019-03-21 3dd6dec
Pixel weights smoothing (#545)
* Support pixel weights in smoothing

* use data.conf.set_temp to set url and timeout
