### 2019-03-21 a6db45f
added healpy_get_wmap_maps.sh script
