### 2019-03-21 af14a6d
Jupyter Notebook tutorial
