+++ doc/tutorial.rst
@@ -1,129 +1,260 @@
-Healpy tutorial
-=================
 
-Creating and manipulating maps
-------------------------------
+healpy tutorial
+===============
 
-Maps are simply numpy arrays, where each array element refers to a location in the sky as defined by the Healpix pixelization schemes (see the `healpix website`_).
+See the Jupyter Notebook version of this tutorial at
+https://github.com/healpy/healpy/blob/master/doc/healpy_tutorial.ipynb
 
-Note: Running the code below in a regular Python session will not display the maps; it's recommended to use an IPython shell or a `Jupyter notebook <https://jupyter.org/install>`_.
-We recommend to use the interactive mode for IPython:
+See a executed version of the notebook with embedded plots at
+https://gist.github.com/zonca/9c114608e0903a3b8ea0bfe41c96f255
 
-.. code-block:: bash
+Choose the ``inline`` backend of ``maptlotlib`` to display the plots
+inside the Jupyter Notebook
 
-    >>> import matplotlib.pyplot as plt
-    >>> plt.ion()
+.. code-block:: python
 
-For Jupyter notebooks, we recommend to switch the backend to ``inline``:
+    import matplotlib.pyplot as plt
+    
+    %matplotlib inline
 
->>> %matplotlib inline
+.. code-block:: python
 
-The resolution of the map is defined by the *NSIDE* parameter. The :py:func:`~healpy.pixelfunc.nside2npix` function gives the number of pixel *NPIX* of the map:
+    import numpy as np
+    import healpy as hp
 
->>> import numpy as np
->>> import healpy as hp
->>> NSIDE = 32
->>> m = np.arange(hp.nside2npix(NSIDE))
->>> hp.mollview(m, title="Mollview image RING")
+NSIDE and ordering
+------------------
 
-.. image:: static/moll_nside32_ring.png
+Maps are simply numpy arrays, where each array element refers to a
+location in the sky as defined by the Healpix pixelization schemes (see
+the `healpix website <https://healpix.jpl.nasa.gov/>`__).
 
-Healpix supports two different ordering schemes, *RING* or *NESTED*. By default, healpy maps are in *RING* ordering.
+Note: Running the code-block below in a regular Python session will not
+display the maps; it’s recommended to use an IPython shell or a Jupyter
+notebook.
 
-In order to work with *NESTED* ordering, all map related functions support the `nest` keyword, for example:
+The resolution of the map is defined by the *NSIDE* parameter, which is
+generally a power of 2.
 
->>> hp.mollview(m, nest=True, title="Mollview image NESTED")
+.. code-block:: python
 
-.. image:: static/moll_nside32_nest.png
+    NSIDE = 32
+    print(
+        "Approximate resolution at NSIDE {} is {:.2} deg".format(
+            NSIDE, hp.nside2resol(NSIDE, arcmin=True) / 60
+        )
+    )
 
-.. _healpix website: http://healpix.sourceforge.net
+The function ``healpy.pixelfunc.nside2npix`` gives the number of pixels
+*NPIX* of the map:
+
+.. code-block:: python
+
+    NPIX = hp.nside2npix(NSIDE)
+    print(NPIX)
+
+The same pixels in the map can be ordered in 2 ways, either RING, where
+they are numbered in the array in horizontal rings starting from the
+North pole:
+
+.. code-block:: python
+
+    m = np.arange(NPIX)
+    hp.mollview(m, title="Mollview image RING")
+    hp.graticule()
+
+The standard coordinates are the **colatitude** :math:`\theta`,
+:math:`0` at the North Pole, :math:`\pi/2` at the equator and
[…]
