### 2019-03-22 #1298 @zonca
ok @harpolea I have released 1.12.9, I renamed the downloader script to `healpy_get_wmap_maps.sh` and install it in the package so it's in the path.

Also, I moved the tutorial to a Jupyter Notebook and now we have a mybinder, see the badge in https://github.com/healpy/healpy

how do I change the version which is attached to the release?
