+++ RELEASE.md
@@ -39,7 +39,7 @@ review and merge the Pull Request at <https://github.com/conda-forge/healpy-feed
 
 ### Linux
 
-    mkdir -p wheelhouse && docker run --rm -v $(pwd)/wheelhouse:/wheelhouse quay.io/pypa/manylinux1_x86_64 bash -c 'for PIP in /opt/python/*/bin/pip; do $PIP install numpy==1.13.3\;python_version\<\"3.7\" numpy==1.14.3\;python_version\>=\"3.7\" && $PIP wheel --no-deps healpy==1.12.7; done; for WHEEL in *.whl; do auditwheel repair $WHEEL; done'
+    mkdir -p wheelhouse && docker run --rm -v $(pwd)/wheelhouse:/wheelhouse quay.io/pypa/manylinux1_x86_64 bash -c 'for PIP in /opt/python/*/bin/pip; do $PIP install numpy==1.13.3\;python_version\<\"3.7\" numpy==1.14.3\;python_version\>=\"3.7\" && $PIP wheel --no-deps healpy==1.12.9; done; for WHEEL in *.whl; do auditwheel repair $WHEEL; done'
 
 ### macOS + MacPorts
 
@@ -49,7 +49,7 @@ review and merge the Pull Request at <https://github.com/conda-forge/healpy-feed
     export CXX=g++-mp-8
     export CFLAGS=-Wa,-q
     export CXXFLAGS=-Wa,-q
-    for VERS in {2.7,3.5,3.6,3.7}; do rm -rf env && virtualenv-$VERS --system-site-packages env && env/bin/pip install --upgrade pip setuptools wheel && env/bin/pip install "numpy==1.13.3;python_version<'3.7'" "numpy==1.14.3;python_version>='3.7'" && env/bin/pip wheel --verbose --no-deps healpy==1.12.7; done
+    for VERS in {2.7,3.5,3.6,3.7}; do rm -rf env && virtualenv-$VERS --system-site-packages env && env/bin/pip install --upgrade pip setuptools wheel && env/bin/pip install "numpy==1.13.3;python_version<'3.7'" "numpy==1.14.3;python_version>='3.7'" && env/bin/pip wheel --verbose --no-deps healpy==1.12.9; done
     python3.7 -m venv --system-site-packages delocate
     delocate/bin/pip install delocate
     for WHEEL in *.whl; do delocate/bin/delocate-wheel -w wheelhouse $WHEEL; done
