### 2019-03-25 cd92337
[paper] mention figure in the text
