### 2019-03-25 #1298 @zonca
thanks @harpolea and @xuanxu, see http://doi.org/10.5281/zenodo.2605426
