+++ paper/paper.md
@@ -48,7 +48,7 @@ Microwave Background, like WMAP [@wmap13] and Planck [@planck18], produce all-sk
 and higher resolution.
 Handling those datasets efficiently and studying their statistical properties
 requires a discretization scheme on the sphere.
-The Hierarchical Equal Area isoLatitude Pixelization ([``HEALPix``](https://healpix.sourceforge.io), [@gorski05]) scheme (see Fig. 1)
+The Hierarchical Equal Area isoLatitude Pixelization, [``HEALPix``](https://healpix.sourceforge.io) [@gorski05], scheme (see Fig. 1)
 has proven to be an excellent mathematical framework to store map-domain data
 and efficiently compute their Spherical Harmonics Transform, whose Angular
 Power Spectrum is one the most powerful tools to understand the early Universe.
@@ -76,9 +76,9 @@ outputs of Cosmic Microwave Background experiments. WMAP, Planck
 and many other experiment in the field release their maps
 in ``HEALPix`` FITS format that can be accessed, visualized, and analyzed with ``healpy``.
 More recently, its usage spread to
-other branches of Astrophysics including galaxy surveys ([Sloan Digital
-Sky Survey depth maps](http://risa.stanford.edu/redmapper/)), Gamma-ray astronomy
-(Fermi [@deil17]), and gravitational waves measurements (LIGO [@singer16]).
+other branches of Astrophysics including galaxy surveys e.g. [Sloan Digital
+Sky Survey depth maps](http://risa.stanford.edu/redmapper/), Gamma-ray astronomy
+e.g. Fermi [@deil17], and gravitational waves measurements e.g. LIGO [@singer16].
 
 # How to cite
 
@@ -89,7 +89,7 @@ and cite this paper and the original ``HEALPix`` paper [@gorski05].
 # Acknowledgements
 
 We acknowledge the Planck collaboration and in particular Peter Meinhold, Julian Borrill and Charles Lawrence
-for supporting Andrea Zonca's work on the project. We thanks all past and current contributors, see
-the [most updated list on Github](https://github.com/healpy/healpy/graphs/contributors).
+for supporting Andrea Zonca's work on the project. We thank all past and current contributors. For a full overview, see
+the [latest list of contributors on Github](https://github.com/healpy/healpy/graphs/contributors).
 
 # References
