### 2019-03-27 #1298 @xuanxu
@zonca: Could you edit the metadata of the Zenodo deposit so it matches the title and author list of the JOSS paper? Thanks!
