### 2019-03-27 #1298 @xuanxu
This is ready now for publication.
Pinging EIC for final acceptance: @openjournals/joss-eics
