### 2019-03-27 #1298 @zonca
Thanks @harpolea, @xuanxu, @danielskatz, it was a great experience publishing with JOSS!
